reformat editor component
